调整结构, 使整个流程更加清晰

+++ src/main/java/org/qiunet/handler/handler/RequestHandlerMapping.java
@@ -60,7 +60,7 @@ public class RequestHandlerMapping {
 	 * @param requestId
 	 * @return
 	 */
-	public IHandler getHandler(int requestId) {
+	public IHandler getHandler(short requestId) {
 		return handlers.get(requestId);
 	}
 }
